+++ package.json
@@ -14,6 +14,7 @@
     "devDependencies": {
         "@types/dotenv": "^8.2.0",
         "@types/node": "^16.11.39",
+        "@types/ws": "^8.5.3",
         "@typescript-eslint/eslint-plugin": "^5.28.0",
         "@typescript-eslint/parser": "^5.28.0",
         "eslint": "^8.17.0",
